+++ noodles-bgzf/examples/bgzf_seek_uncompressed.rs
@@ -19,7 +19,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     let position = args.next().expect("missing position").parse()?;
     let length = args.next().expect("missing length").parse()?;
 
-    let mut reader = bgzf::indexed_reader::Builder::default().build_from_path(src)?;
+    let mut reader = bgzf::io::indexed_reader::Builder::default().build_from_path(src)?;
     reader.seek(SeekFrom::Start(position))?;
 
     let mut buf = vec![0; length];