+++ noodles-bgzf/src/io/indexed_reader.rs
@@ -6,7 +6,8 @@ pub use self::builder::Builder;
 
 use std::io::{self, BufRead, Read, Seek, SeekFrom};
 
-use super::{gzi, io::Reader, VirtualPosition};
+use super::Reader;
+use crate::{gzi, VirtualPosition};
 
 /// An indexed BGZF reader.
 pub struct IndexedReader<R> {